feat: 添加快速配置读写上下文的扩展

+++ Demo/EntityLib/SecondEntity.cs
@@ -5,7 +5,7 @@ namespace DataDemo.EntityLib;
 /// <summary>
 /// 实体2
 /// </summary>
-public class SecondEntity : BaseEntity<long?>
+public class SecondEntity : BaseEntity<int?>
 {
     public string Name { get; set; }
     public int? Age { get; set; }
